fix: remove ai-chat-utils
private package was breaking stackblitz
