+++ demo/tests/utils.ts
@@ -6,11 +6,16 @@
  */
 
 import { PageObjectId, ViewType } from "@carbon/ai-chat/server";
-import { sleep } from "@carbon/ai-chat-utils";
 import type { Page } from "@playwright/test";
 
 // Import types for window globals used in evaluated browser context.
 import type {} from "../types/window";
+
+async function sleep(milliseconds: number) {
+  await new Promise((resolve) => {
+    setTimeout(resolve, milliseconds);
+  });
+}
 
 /**
  * Demo-specific test IDs that are not part of the core @carbon/ai-chat package.
