+++ examples/react/custom-element/src/customSendMessage.ts
@@ -14,7 +14,12 @@ import {
   MessageResponseTypes,
   StreamChunk,
 } from "@carbon/ai-chat";
-import { sleep } from "@carbon/ai-chat-utils";
+
+async function sleep(milliseconds: number) {
+  await new Promise((resolve) => {
+    setTimeout(resolve, milliseconds);
+  });
+}
 
 const WELCOME_TEXT = `Welcome to this example of a custom back-end. This back-end is mocked entirely on the client side. It does not show all potential functionality.
