+++ packages/ai-chat/package.json
@@ -83,7 +83,6 @@
     "@types/node": "^24.0.12",
     "@types/react": "^19.2.1",
     "@types/react-dom": "^19.2.1",
-    "@carbon/ai-chat-utils": "0.0.0",
     "@open-wc/testing": "4.0.0",
     "autoprefixer": "^10.4.21",
     "babel-jest": "^29.7.0",
