--- utils/index.d.ts
@@ -1,19 +1,0 @@
-/*
- *  Copyright IBM Corp. 2025
- *
- *  This source code is licensed under the Apache-2.0 license found in the
- *  LICENSE file in the root directory of this source tree.
- *
- *  @license
- */
-
-export declare const TABLE: string;
-export declare const UNORDERED_LIST: string;
-export declare const ORDERED_LIST: string;
-export declare const TEXT: string;
-export declare const HEADERS: string;
-export declare const CODE: string;
-export declare const BLOCKQUOTE: string;
-export declare const MARKDOWN: string;
-export declare const HTML: string;
-export declare function sleep(milliseconds: number): Promise<void>;
